Problem:  hg results were different between linux, stratus, paunchy. Solutions: The linux machines silently ignored floating point overflow in the scaling routines. I modified the hg routines to use double precision instead of single precision almost everywhere. Stratus apparently promotes long ints to 64 bits even without the -m64 compilation flag.  This caused the Zoltan "universal" random number generator to produce different numbers.  This has been fixed (assuming that no target machine uses 16 bit ints as the default) locally in hg. The testscript was modified slightly to not assume the search path contains the current directory. There is still a very tiny difference (7 of the 414 test variants) between Stratus and the other platforms.  I suspect tie breaking differences.  This is still under investigation.  These results are the new reference set and differ only slightly from the previous reference set.

+++ packages/zoltan/src/hg/hg_global.c
@@ -584,7 +584,7 @@ static int greedy_order (
   int esize, *vtx_count=NULL, *visited=NULL, *cut[2];
   int ierr=ZOLTAN_OK;
   double weight_sum= 0.0, part_sum= 0.0, old_sum, cutoff;
-  float *gain = NULL, *edge_sum = NULL, delta;
+  double *gain = NULL, *edge_sum = NULL, delta;
   double damp_factor;
   char msg[128];
   HEAP h[2];
@@ -595,7 +595,7 @@ static int greedy_order (
 
   /* Allocate arrays. */
   if (!(rank  = (int *)     ZOLTAN_CALLOC (hg->nVtx, sizeof (int))) ||
-      !(gain  = (float *)   ZOLTAN_CALLOC (hg->nVtx, sizeof (float))) ) {
+      !(gain  = (double *)   ZOLTAN_CALLOC (hg->nVtx, sizeof (double))) ) {
     ZOLTAN_PRINT_ERROR(zz->Proc, yo, "Insufficient memory.");
     ierr =  ZOLTAN_MEMERR;
     goto error;
@@ -604,7 +604,7 @@ static int greedy_order (
     rank[i] = -1;  /* -1 means this vtx has not yet been numbered */
 
   if (priority_mode && (!(priority_mode&1))){ /* 2,4,6,... */
-    if (!(edge_sum = (float *)  ZOLTAN_CALLOC (hg->nVtx, sizeof (float)))){
+    if (!(edge_sum = (double *)  ZOLTAN_CALLOC (hg->nVtx, sizeof (double)))){
       ZOLTAN_PRINT_ERROR(zz->Proc, yo, "Insufficient memory.");
       ierr =  ZOLTAN_MEMERR;
       goto error;
